reorganized files in EoP
