+++ active_projects/eop/chapter1/intro.py
@@ -1,0 +1,55 @@
+from big_ol_pile_of_manim_imports import *
+from active_projects.eop.reusables import *
+
+class Chapter1OpeningQuote(OpeningQuote):
+    CONFIG = {
+        "fade_in_kwargs": {
+            "submobject_mode": "lagged_start",
+            "rate_func": None,
+            "lag_factor": 9,
+            "run_time": 10,
+        },
+        "text_size" : "\\normalsize",
+        "use_quotation_marks": False,
+        "quote" : [
+            "To see a world in a grain of sand\\\\",
+            "And a heaven in a wild flower,\\\\",
+            "Hold infinity in the palm of your hand\\\\",
+            "\phantom{r}And eternity in an hour.\\\\"
+        ],
+        "quote_arg_separator" : " ",
+        "highlighted_quote_terms" : {},
+        "author" : "William Blake: \\\\ \emph{Auguries of Innocence}",
+    }
+
+class Introduction(TeacherStudentsScene):
+
+    CONFIG = {
+        "default_pi_creature_kwargs": {
+        "color": MAROON_E,
+        "flip_at_start": True,
+        },
+    }
+
+    def construct(self):
+        
+        self.wait(5)
+        
+        self.change_student_modes(
+            "confused", "frustrated", "dejected",
+            look_at_arg = UP + 2 * RIGHT
+        )
+
+        self.wait()
+
+
+        self.get_teacher().change_mode("raise_right_hand")
+        self.wait()
+
+        self.wait(30)
+        # put examples here in video editor
+
+
+        # # # # # # # # # # # # # # # # # #
+        # show examples of the area model #
+        # # # # # # # # # # # # # # # # # #
