+++ active_projects/eop/chapter2/permutation_grid.py
