+++ active_projects/eop/reusables/eop_constants.py
@@ -1,0 +1,18 @@
+
+from constants import *
+
+COIN_RADIUS = 0.18
+COIN_THICKNESS = 0.4 * COIN_RADIUS
+COIN_FORESHORTENING = 0.5
+COIN_NB_RIDGES = 20
+COIN_STROKE_WIDTH = 2
+
+COIN_SEQUENCE_SPACING = 0.1
+
+GRADE_COLOR_1 = COLOR_HEADS = RED
+GRADE_COLOR_2 = COLOR_TAILS = BLUE_E
+
+TALLY_BACKGROUND_WIDTH = 1.0
+TALLY_BACKGROUND_COLOR = BLACK
+
+SICKLY_GREEN = "#9BBD37"
