+++ active_projects/eop/reusables/histograms.py
@@ -280,6 +280,49 @@ class FlashThroughHistogram(Animation):
 
 
 
+class OutlineableBars(VGroup):
+
+    # A group of bars (rectangles), together with
+    # a method that draws an outline around them,
+    # assuming the bars are arranged in a histogram
+    # (aligned at the bottom without gaps).
+
+    # We use this to morph a row of bricks into a histogram.
+
+    CONFIG = {
+        "outline_stroke_width" : 5,
+        "stroke_color" : WHITE
+    }
+    def create_outline(self, animated = False, **kwargs):
+
+        outline_points = []
+
+        for (i, bar) in enumerate(self.submobjects):
+            
+            if i == 0:
+                # start with the lower left
+                outline_points.append(bar.get_corner(DOWN + LEFT))
+
+            # upper two points of each bar
+            outline_points.append(bar.get_corner(UP + LEFT))
+            outline_points.append(bar.get_corner(UP + RIGHT))
+
+            previous_bar = bar
+        # close the outline
+            # lower right
+        outline_points.append(previous_bar.get_corner(DOWN + RIGHT))
+            # lower left
+        outline_points.append(outline_points[0])
+
+        self.outline = Polygon(*outline_points,
+            stroke_width = self.outline_stroke_width,
+            stroke_color = self.stroke_color)
+
+        if animated:
+            self.play(FadeIn(self.outline, **kwargs))
+        return self.outline
+
+
